requirements.txt
gym==0.10.5
mujoco-py==1.50.1.56
tensorflow-gpu
numpy==1.14.5
